refactor(generator): generators better orginized with bugfixes
