+++ .run/Run Gemini Read.run.xml
@@ -3,7 +3,7 @@
     <module name="gemini" />
     <working_directory value="$PROJECT_DIR$" />
     <go_parameters value="-gcflags &quot;all=-N -l&quot;" />
-    <parameters value="--dataset-size=small --cql-features all --warmup 0 --duration 2m --drop-schema true --fail-fast --level info --non-interactive --materialized-views false --outfile ./results/gemini.log --test-statement-log-file ./results/gemini_test_statements.log --oracle-statement-log-file ./results/gemini_oracle_statements.log --test-host-selection-policy token-aware --oracle-host-selection-policy token-aware --test-cluster=192.168.100.2 --oracle-cluster=192.168.100.3 --outfile ./results/gemini_result.log --mode read --non-interactive --request-timeout 180s --connect-timeout 120s --use-server-timestamps false --async-objects-stabilization-attempts 10 --async-objects-stabilization-backoff 100ms --replication-strategy &quot;{'class': 'NetworkTopologyStrategy', 'replication_factor': '1'}&quot; --oracle-replication-strategy &quot;{'class': 'NetworkTopologyStrategy', 'replication_factor': '1'}&quot; --max-mutation-retries 5 --max-mutation-retries-backoff 1000ms --concurrency 2 --tracing-outfile ./results/gemini_tracing.log" />
+    <parameters value="--mode read --seed=100 --schema-seed=100 --dataset-size=small --cql-features normal --warmup 0 --duration 2m --drop-schema true --fail-fast --level info --test-host-selection-policy token-aware --oracle-host-selection-policy token-aware --test-cluster=192.168.100.2 --oracle-cluster=192.168.100.3 --non-interactive --request-timeout 180s --connect-timeout 120s --use-server-timestamps false --async-objects-stabilization-attempts 10 --async-objects-stabilization-backoff 100ms --replication-strategy &quot;{'class': 'NetworkTopologyStrategy', 'replication_factor': '1'}&quot; --oracle-replication-strategy &quot;{'class': 'NetworkTopologyStrategy', 'replication_factor': '1'}&quot; --max-mutation-retries 5 --max-mutation-retries-backoff 1000ms --concurrency 1 --profiling-port 6060" />
     <kind value="PACKAGE" />
     <package value="github.com/scylladb/gemini/cmd/gemini" />
     <directory value="$PROJECT_DIR$" />
