+++ cmd/gemini/root.go
@@ -15,12 +15,7 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"log"
-	"math"
-	"net/http"
-	"net/http/pprof"
 	"os"
 	"os/signal"
 	"strconv"
@@ -30,11 +25,9 @@ import (
 	"time"
 
 	"github.com/scylladb/gemini/pkg/auth"
-	"github.com/scylladb/gemini/pkg/builders"
 	"github.com/scylladb/gemini/pkg/generators"
 	"github.com/scylladb/gemini/pkg/jobs"
 	"github.com/scylladb/gemini/pkg/realrandom"
-	"github.com/scylladb/gemini/pkg/replication"
 	"github.com/scylladb/gemini/pkg/store"
 	"github.com/scylladb/gemini/pkg/typedef"
 	"github.com/scylladb/gemini/pkg/utils"
@@ -44,97 +37,15 @@ import (
 	"github.com/gocql/gocql"
 	"github.com/hailocab/go-hostpool"
 	"github.com/pkg/errors"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"golang.org/x/exp/rand"
 	"golang.org/x/net/context"
-	"gonum.org/v1/gonum/stat/distuv"
-)
-
-var (
-	testClusterHost                  []string
-	testClusterUsername              string
-	testClusterPassword              string
-	oracleClusterHost                []string
-	oracleClusterUsername            string
-	oracleClusterPassword            string
-	schemaFile                       string
-	outFileArg                       string
-	concurrency                      uint64
-	seed                             string
-	schemaSeed                       string
-	dropSchema                       bool
-	verbose                          bool
-	mode                             string
-	failFast                         bool
-	nonInteractive                   bool
-	duration                         time.Duration
-	bind                             string
-	warmup                           time.Duration
-	replicationStrategy              string
-	tableOptions                     []string
-	oracleReplicationStrategy        string
-	consistency                      string
-	maxTables                        int
-	maxPartitionKeys                 int
-	minPartitionKeys                 int
-	maxClusteringKeys                int
-	minClusteringKeys                int
-	maxColumns                       int
-	minColumns                       int
-	datasetSize                      string
-	cqlFeatures                      string
-	useMaterializedViews             bool
-	level                            string
-	maxRetriesMutate                 int
-	maxRetriesMutateSleep            time.Duration
-	maxErrorsToStore                 int
-	pkBufferReuseSize                uint64
-	partitionCount                   uint64
-	partitionKeyDistribution         string
-	normalDistMean                   float64
-	normalDistSigma                  float64
-	tracingOutFile                   string
-	useCounters                      bool
-	asyncObjectStabilizationAttempts int
-	asyncObjectStabilizationDelay    time.Duration
-	useLWT                           bool
-	testClusterHostSelectionPolicy   string
-	oracleClusterHostSelectionPolicy string
-	useServerSideTimestamps          bool
-	requestTimeout                   time.Duration
-	connectTimeout                   time.Duration
-	profilingPort                    int
-	testStatementLogFile             string
-	oracleStatementLogFile           string
 )
 
 func interactive() bool {
 	return !nonInteractive
-}
-
-func readSchema(confFile string, schemaConfig typedef.SchemaConfig) (*typedef.Schema, error) {
-	byteValue, err := os.ReadFile(confFile)
-	if err != nil {
-		return nil, err
-	}
-
-	var shm typedef.Schema
-
-	err = json.Unmarshal(byteValue, &shm)
-	if err != nil {
-		return nil, err
-	}
-
-	schemaBuilder := builders.NewSchemaBuilder()
-	schemaBuilder.Keyspace(shm.Keyspace).Config(schemaConfig)
-	for t, tbl := range shm.Tables {
-		shm.Tables[t].LinkIndexAndColumns()
-		schemaBuilder.Table(tbl)
-	}
-	return schemaBuilder.Build(), nil
 }
 
 func run(_ *cobra.Command, _ []string) error {
@@ -154,11 +65,7 @@ func run(_ *cobra.Command, _ []string) error {
 
 	rand.Seed(intSeed)
 
-	cons, err := gocql.ParseConsistencyWrapper(consistency)
-	if err != nil {
-		logger.Error("Unable parse consistency, error=%s. Falling back on Quorum", zap.Error(err))
-		cons = gocql.Quorum
-	}
+	cons := gocql.ParseConsistency(consistency)
 
 	testHostSelectionPolicy, err := getHostSelectionPolicy(testClusterHostSelectionPolicy, testClusterHost)
 	if err != nil {
@@ -172,18 +79,23 @@ func run(_ *cobra.Command, _ []string) error {
 	ctx, cancel := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGABRT, syscall.SIGTERM)
 	defer cancel()
 
-	go func() {
-		http.Handle("/metrics", promhttp.Handler())
-		_ = http.ListenAndServe(bind, nil)
-	}()
-
-	if profilingPort != 0 {
-		go func() {
-			mux := http.NewServeMux()
-			mux.HandleFunc("/profile", pprof.Profile)
-			log.Fatal(http.ListenAndServe(":"+strconv.Itoa(profilingPort), mux))
-		}()
-	}
+	//go func() {
+	//	mux := http.NewServeMux()
+	//	mux.Handle("GET /metrics", promhttp.Handler())
+	//	log.Fatal(http.ListenAndServe(bind, mux))
+	//}()
+	//
+	//if profilingPort != 0 {
+	//	go func() {
+	//		mux := http.NewServeMux()
+	//		mux.HandleFunc("GET /debug/pprof/", pprof.Index)
+	//		mux.HandleFunc("GET /debug/pprof/cmdline", pprof.Cmdline)
+	//		mux.HandleFunc("GET /debug/pprof/profile", pprof.Profile)
+	//		mux.HandleFunc("GET /debug/pprof/symbol", pprof.Symbol)
+	//		mux.HandleFunc("GET /debug/pprof/trace", pprof.Trace)
+	//		log.Fatal(http.ListenAndServe(":"+strconv.Itoa(profilingPort), mux))
+	//	}()
+	//}
 
 	outFile, err := createFile(outFileArg, os.Stdout)
 	if err != nil {
@@ -208,10 +120,7 @@ func run(_ *cobra.Command, _ []string) error {
 		}
 	}
 
-	jsonSchema, _ := json.MarshalIndent(schema, "", "    ")
-
-	printSetup(intSeed, intSchemaSeed)
-	fmt.Printf("Schema: %v\n", string(jsonSchema))
+	printSetup(intSeed, intSchemaSeed, mode)
 
 	testCluster, oracleCluster := createClusters(cons, testHostSelectionPolicy, oracleHostSelectionPolicy, logger)
 	storeConfig := store.Config{
@@ -234,14 +143,17 @@ func run(_ *cobra.Command, _ []string) error {
 				return ioErr
 			}
 			tracingFile = tf
-			defer utils.IgnoreError(tracingFile.Sync)
+			defer func() {
+				utils.IgnoreError(tracingFile.Sync)
+				utils.IgnoreError(tracingFile.Close)
+			}()
 		}
 	}
 	st, err := store.New(schema, testCluster, oracleCluster, storeConfig, tracingFile, logger)
 	if err != nil {
 		return err
 	}
-	defer utils.IgnoreError(st.Close)
+	//defer utils.IgnoreError(st.Close)
 
 	if dropSchema && mode != jobs.ReadMode {
 		for _, stmt := range generators.GetDropKeyspace(schema) {
@@ -254,34 +166,41 @@ func run(_ *cobra.Command, _ []string) error {
 
 	testKeyspace, oracleKeyspace := generators.GetCreateKeyspaces(schema)
 	if err = st.Create(
-		context.Background(),
+		ctx,
 		typedef.SimpleStmt(testKeyspace, typedef.CreateKeyspaceStatementType),
 		typedef.SimpleStmt(oracleKeyspace, typedef.CreateKeyspaceStatementType)); err != nil {
 		return errors.Wrap(err, "unable to create keyspace")
 	}
 
-	for _, stmt := range generators.GetCreateSchema(schema) {
+	for _, stmt := range generators.GetCreateSchema(schema, useMaterializedViews) {
 		logger.Debug(stmt)
 		if err = st.Mutate(ctx, typedef.SimpleStmt(stmt, typedef.CreateSchemaStatementType)); err != nil {
 			return errors.Wrap(err, "unable to create schema")
 		}
 	}
 
-	ctx, done := context.WithTimeout(ctx, duration+warmup+10*time.Second)
-
-	gens, err := createGenerators(schema, schemaConfig, intSeed, partitionCount, logger)
+	distFunc, err := generators.ParseDistributionDefault(partitionKeyDistribution, partitionCount, intSeed)
 	if err != nil {
 		return err
 	}
-	gens.StartAll(ctx)
+
+	gens, err := generators.New(ctx, schema, schemaConfig, intSeed, partitionCount, logger, distFunc, pkBufferReuseSize)
+	if err != nil {
+		return err
+	}
+
+	defer utils.IgnoreError(gens.Close)
+
+	ctx, done := context.WithTimeout(ctx, duration+warmup+10*time.Second)
+	defer done()
 
 	if !nonInteractive {
 		sp := createSpinner(interactive())
 		ticker := time.NewTicker(time.Second)
 
 		go func() {
-			defer done()
 			defer ticker.Stop()
+
 			for {
 				select {
 				case <-ctx.Done():
@@ -293,9 +212,9 @@ func run(_ *cobra.Command, _ []string) error {
 		}()
 	}
 
-	jobsList := jobs.New(mode, duration, concurrency, logger, schema, st, globalStatus, intSeed, gens, failFast, warmup)
+	jobsList := jobs.New(mode, duration, concurrency, logger, schema, st, globalStatus, intSeed, gens, warmup, failFast)
 	if err = jobsList.Run(ctx); err != nil {
-		logger.Debug("error detected", zap.Error(err))
+		logger.Error("error detected", zap.Error(err))
 	}
 
 	logger.Info("test finished")
@@ -319,37 +238,6 @@ func createFile(name string, def *os.File) (*os.File, error) {
 	}
 
 	return def, nil
-}
-
-const (
-	stdDistMean = math.MaxUint64 / 2
-	oneStdDev   = 0.341 * math.MaxUint64
-)
-
-func createDistributionFunc(distribution string, size, seed uint64, mu, sigma float64) (generators.DistributionFunc, error) {
-	switch strings.ToLower(distribution) {
-	case "zipf":
-		dist := rand.NewZipf(rand.New(rand.NewSource(seed)), 1.1, 1.1, size)
-		return func() generators.TokenIndex {
-			return generators.TokenIndex(dist.Uint64())
-		}, nil
-	case "normal":
-		dist := distuv.Normal{
-			Src:   rand.NewSource(seed),
-			Mu:    mu,
-			Sigma: sigma,
-		}
-		return func() generators.TokenIndex {
-			return generators.TokenIndex(dist.Rand())
-		}, nil
-	case "uniform":
-		rnd := rand.New(rand.NewSource(seed))
-		return func() generators.TokenIndex {
-			return generators.TokenIndex(rnd.Uint64n(size))
-		}, nil
-	default:
-		return nil, errors.Errorf("unsupported distribution: %s", distribution)
-	}
 }
 
 func createLogger(level string) *zap.Logger {
@@ -404,35 +292,8 @@ func createClusters(
 	return testCluster, oracleCluster
 }
 
-func getReplicationStrategy(rs string, fallback *replication.Replication, logger *zap.Logger) *replication.Replication {
-	switch rs {
-	case "network":
-		return replication.NewNetworkTopologyStrategy()
-	case "simple":
-		return replication.NewSimpleStrategy()
-	default:
-		replicationStrategy := &replication.Replication{}
-		if err := json.Unmarshal([]byte(strings.ReplaceAll(rs, "'", "\"")), replicationStrategy); err != nil {
-			logger.Error("unable to parse replication strategy", zap.String("strategy", rs), zap.Error(err))
-			return fallback
-		}
-		return replicationStrategy
-	}
-}
-
-func getCQLFeature(feature string) typedef.CQLFeature {
-	switch strings.ToLower(feature) {
-	case "all":
-		return typedef.CQL_FEATURE_ALL
-	case "normal":
-		return typedef.CQL_FEATURE_NORMAL
-	default:
-		return typedef.CQL_FEATURE_BASIC
-	}
-}
-
 func getHostSelectionPolicy(policy string, hosts []string) (gocql.HostSelectionPolicy, error) {
-	switch policy {
+	switch strings.ToLower(policy) {
 	case "round-robin":
 		return gocql.RoundRobinHostPolicy(), nil
 	case "host-pool":
@@ -451,89 +312,7 @@ var rootCmd = &cobra.Command{
 	SilenceUsage: true,
 }
 
-func init() {
-	rootCmd.Version = version + ", commit " + commit + ", date " + date
-	rootCmd.Flags().StringSliceVarP(&testClusterHost, "test-cluster", "t", []string{}, "Host names or IPs of the test cluster that is system under test")
-	_ = rootCmd.MarkFlagRequired("test-cluster")
-	rootCmd.Flags().StringVarP(&testClusterUsername, "test-username", "", "", "Username for the test cluster")
-	rootCmd.Flags().StringVarP(&testClusterPassword, "test-password", "", "", "Password for the test cluster")
-	rootCmd.Flags().StringSliceVarP(
-		&oracleClusterHost, "oracle-cluster", "o", []string{},
-		"Host names or IPs of the oracle cluster that provides correct answers. If omitted no oracle will be used")
-	rootCmd.Flags().StringVarP(&oracleClusterUsername, "oracle-username", "", "", "Username for the oracle cluster")
-	rootCmd.Flags().StringVarP(&oracleClusterPassword, "oracle-password", "", "", "Password for the oracle cluster")
-	rootCmd.Flags().StringVarP(&schemaFile, "schema", "", "", "Schema JSON config file")
-	rootCmd.Flags().StringVarP(&mode, "mode", "m", jobs.MixedMode, "Query operation mode. Mode options: write, read, mixed (default)")
-	rootCmd.Flags().Uint64VarP(&concurrency, "concurrency", "c", 10, "Number of threads per table to run concurrently")
-	rootCmd.Flags().StringVarP(&seed, "seed", "s", "random", "Statement seed value")
-	rootCmd.Flags().StringVarP(&schemaSeed, "schema-seed", "", "random", "Schema seed value")
-	rootCmd.Flags().BoolVarP(&dropSchema, "drop-schema", "d", false, "Drop schema before starting tests run")
-	rootCmd.Flags().BoolVarP(&verbose, "verbose", "v", false, "Verbose output during test run")
-	rootCmd.Flags().BoolVarP(&failFast, "fail-fast", "f", false, "Stop on the first failure")
-	rootCmd.Flags().BoolVarP(&nonInteractive, "non-interactive", "", false, "Run in non-interactive mode (disable progress indicator)")
-	rootCmd.Flags().DurationVarP(&duration, "duration", "", 30*time.Second, "")
-	rootCmd.Flags().StringVarP(&outFileArg, "outfile", "", "", "Specify the name of the file where the results should go")
-	rootCmd.Flags().StringVarP(&bind, "bind", "b", ":2112", "Specify the interface and port which to bind prometheus metrics on. Default is ':2112'")
-	rootCmd.Flags().DurationVarP(&warmup, "warmup", "", 30*time.Second, "Specify the warmup perid as a duration for example 30s or 10h")
-	rootCmd.Flags().StringVarP(
-		&replicationStrategy, "replication-strategy", "", "simple",
-		"Specify the desired replication strategy as either the coded short hand simple|network to get the default for each type or provide "+
-			"the entire specification in the form {'class':'....'}")
-	rootCmd.Flags().StringVarP(
-		&oracleReplicationStrategy, "oracle-replication-strategy", "", "simple",
-		"Specify the desired replication strategy of the oracle cluster as either the coded short hand simple|network to get the default for each "+
-			"type or provide the entire specification in the form {'class':'....'}")
-	rootCmd.Flags().StringArrayVarP(&tableOptions, "table-options", "", []string{}, "Repeatable argument to set table options to be added to the created tables")
-	rootCmd.Flags().StringVarP(&consistency, "consistency", "", "LOCAL_QUORUM", "Specify the desired consistency as ANY|ONE|TWO|THREE|QUORUM|LOCAL_QUORUM|EACH_QUORUM|LOCAL_ONE")
-	rootCmd.Flags().IntVarP(&maxTables, "max-tables", "", 1, "Maximum number of generated tables")
-	rootCmd.Flags().IntVarP(&maxPartitionKeys, "max-partition-keys", "", 6, "Maximum number of generated partition keys")
-	rootCmd.Flags().IntVarP(&minPartitionKeys, "min-partition-keys", "", 2, "Minimum number of generated partition keys")
-	rootCmd.Flags().IntVarP(&maxClusteringKeys, "max-clustering-keys", "", 4, "Maximum number of generated clustering keys")
-	rootCmd.Flags().IntVarP(&minClusteringKeys, "min-clustering-keys", "", 2, "Minimum number of generated clustering keys")
-	rootCmd.Flags().IntVarP(&maxColumns, "max-columns", "", 16, "Maximum number of generated columns")
-	rootCmd.Flags().IntVarP(&minColumns, "min-columns", "", 8, "Minimum number of generated columns")
-	rootCmd.Flags().StringVarP(&datasetSize, "dataset-size", "", "large", "Specify the type of dataset size to use, small|large")
-	rootCmd.Flags().StringVarP(&cqlFeatures, "cql-features", "", "basic", "Specify the type of cql features to use, basic|normal|all")
-	rootCmd.Flags().BoolVarP(&useMaterializedViews, "materialized-views", "", false, "Run gemini with materialized views support")
-	rootCmd.Flags().StringVarP(&level, "level", "", "info", "Specify the logging level, debug|info|warn|error|dpanic|panic|fatal")
-	rootCmd.Flags().IntVarP(&maxRetriesMutate, "max-mutation-retries", "", 2, "Maximum number of attempts to apply a mutation")
-	rootCmd.Flags().DurationVarP(
-		&maxRetriesMutateSleep, "max-mutation-retries-backoff", "", 10*time.Millisecond,
-		"Duration between attempts to apply a mutation for example 10ms or 1s")
-	rootCmd.Flags().Uint64VarP(&pkBufferReuseSize, "partition-key-buffer-reuse-size", "", 100, "Number of reused buffered partition keys")
-	rootCmd.Flags().Uint64VarP(&partitionCount, "token-range-slices", "", 10000, "Number of slices to divide the token space into")
-	rootCmd.Flags().StringVarP(
-		&partitionKeyDistribution, "partition-key-distribution", "", "uniform",
-		"Specify the distribution from which to draw partition keys, supported values are currently uniform|normal|zipf")
-	rootCmd.Flags().Float64VarP(&normalDistMean, "normal-dist-mean", "", stdDistMean, "Mean of the normal distribution")
-	rootCmd.Flags().Float64VarP(&normalDistSigma, "normal-dist-sigma", "", oneStdDev, "Sigma of the normal distribution, defaults to one standard deviation ~0.341")
-	rootCmd.Flags().StringVarP(
-		&tracingOutFile, "tracing-outfile", "", "",
-		"Specify the file to which tracing information gets written. Two magic names are available, 'stdout' and 'stderr'. By default tracing is disabled.")
-	rootCmd.Flags().BoolVarP(&useCounters, "use-counters", "", false, "Ensure that at least one table is a counter table")
-	rootCmd.Flags().IntVarP(
-		&asyncObjectStabilizationAttempts, "async-objects-stabilization-attempts", "", 10,
-		"Maximum number of attempts to validate result sets from MV and SI")
-	rootCmd.Flags().DurationVarP(
-		&asyncObjectStabilizationDelay, "async-objects-stabilization-backoff", "", 10*time.Millisecond,
-		"Duration between attempts to validate result sets from MV and SI for example 10ms or 1s")
-	rootCmd.Flags().BoolVarP(&useLWT, "use-lwt", "", false, "Emit LWT based updates")
-	rootCmd.Flags().StringVarP(
-		&oracleClusterHostSelectionPolicy, "oracle-host-selection-policy", "", "token-aware",
-		"Host selection policy used by the driver for the oracle cluster: round-robin|host-pool|token-aware")
-	rootCmd.Flags().StringVarP(
-		&testClusterHostSelectionPolicy, "test-host-selection-policy", "", "token-aware",
-		"Host selection policy used by the driver for the test cluster: round-robin|host-pool|token-aware")
-	rootCmd.Flags().BoolVarP(&useServerSideTimestamps, "use-server-timestamps", "", false, "Use server-side generated timestamps for writes")
-	rootCmd.Flags().DurationVarP(&requestTimeout, "request-timeout", "", 30*time.Second, "Duration of waiting request execution")
-	rootCmd.Flags().DurationVarP(&connectTimeout, "connect-timeout", "", 30*time.Second, "Duration of waiting connection established")
-	rootCmd.Flags().IntVarP(&profilingPort, "profiling-port", "", 0, "If non-zero starts pprof profiler on given port at 'http://0.0.0.0:<port>/profile'")
-	rootCmd.Flags().IntVarP(&maxErrorsToStore, "max-errors-to-store", "", 1000, "Maximum number of errors to store and output at the end")
-	rootCmd.Flags().StringVarP(&testStatementLogFile, "test-statement-log-file", "", "", "File to write statements flow to")
-	rootCmd.Flags().StringVarP(&oracleStatementLogFile, "oracle-statement-log-file", "", "", "File to write statements flow to")
-}
-
-func printSetup(seed, schemaSeed uint64) {
+func printSetup(seed, schemaSeed uint64, mode string) {
 	tw := new(tabwriter.Writer)
 	tw.Init(os.Stdout, 0, 8, 2, '\t', tabwriter.AlignRight)
 	fmt.Fprintf(tw, "Seed:\t%d\n", seed)
@@ -543,6 +322,7 @@ func printSetup(seed, schemaSeed uint64) {
 	fmt.Fprintf(tw, "Concurrency:\t%d\n", concurrency)
 	fmt.Fprintf(tw, "Test cluster:\t%s\n", testClusterHost)
 	fmt.Fprintf(tw, "Oracle cluster:\t%s\n", oracleClusterHost)
+	fmt.Fprintf(tw, "Mode:\t%s\n", mode)
 	if outFileArg == "" {
 		fmt.Fprintf(tw, "Output file:\t%s\n", "<stdout>")
 	} else {
