+++ pkg/replication/replication_test.go
@@ -17,6 +17,8 @@ package replication_test
 import (
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
+
 	"github.com/scylladb/gemini/pkg/replication"
 )
 
@@ -42,6 +44,39 @@ func TestToCQL(t *testing.T) {
 			got := test.rs.ToCQL()
 			if got != test.want {
 				t.Fatalf("expected '%s', got '%s'", test.want, got)
+			}
+		})
+	}
+}
+
+func TestGetReplicationStrategy(t *testing.T) {
+	tests := map[string]struct {
+		strategy string
+		expected string
+	}{
+		"simple strategy": {
+			strategy: "{\"class\": \"SimpleStrategy\", \"replication_factor\": \"1\"}",
+			expected: "{'class':'SimpleStrategy','replication_factor':'1'}",
+		},
+		"simple strategy single quotes": {
+			strategy: "{'class': 'SimpleStrategy', 'replication_factor': '1'}",
+			expected: "{'class':'SimpleStrategy','replication_factor':'1'}",
+		},
+		"network topology strategy": {
+			strategy: "{\"class\": \"NetworkTopologyStrategy\", \"dc1\": 3, \"dc2\": 3}",
+			expected: "{'class':'NetworkTopologyStrategy','dc1':3,'dc2':3}",
+		},
+		"network topology strategy single quotes": {
+			strategy: "{'class': 'NetworkTopologyStrategy', 'dc1': 3, 'dc2': 3}",
+			expected: "{'class':'NetworkTopologyStrategy','dc1':3,'dc2':3}",
+		},
+	}
+
+	for name, tc := range tests {
+		t.Run(name, func(t *testing.T) {
+			got := replication.MustParseReplication(tc.strategy)
+			if diff := cmp.Diff(got.ToCQL(), tc.expected); diff != "" {
+				t.Errorf("expected=%s, got=%s,diff=%s", tc.strategy, got.ToCQL(), diff)
 			}
 		})
 	}
